# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/EventEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/EmployeeRepository.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/EventRepository.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/EmployeeService.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/EmployerService.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/EmployeeServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/EmployerServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/MailServiceImpl.java
#	frontend/src/app/components/event-overview/event-overview.component.html
#	frontend/src/app/components/event-overview/event-overview.component.ts
#	frontend/src/app/services/employer.service.ts

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/EventEndpoint.java
@@ -1,16 +1,23 @@
 package at.ac.tuwien.sepm.groupphase.backend.endpoint;
 
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.DetailedEventDto;
+import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.EmployerDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.EventInquiryDto;
+import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.ProfileDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.SearchEventDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.mapper.EventMapper;
+import at.ac.tuwien.sepm.groupphase.backend.entity.Employer;
+import at.ac.tuwien.sepm.groupphase.backend.entity.Event;
+import at.ac.tuwien.sepm.groupphase.backend.entity.Profile;
 import at.ac.tuwien.sepm.groupphase.backend.service.EventService;
+import at.ac.tuwien.sepm.groupphase.backend.service.TokenService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.Authorization;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -28,11 +35,13 @@ public class EventEndpoint {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private final EventService eventService;
     private final EventMapper eventMapper;
+    private final TokenService tokenService;
 
     @Autowired
-    public EventEndpoint(EventMapper eventMapper, EventService eventService) {
+    public EventEndpoint(EventMapper eventMapper, EventService eventService, TokenService tokenService) {
         this.eventService = eventService;
         this.eventMapper = eventMapper;
+        this.tokenService = tokenService;
     }
 
     @PostMapping
@@ -42,7 +51,8 @@ public class EventEndpoint {
     @PreAuthorize("hasAuthority('ROLE_EMPLOYER')")
     public EventInquiryDto create(@Valid @RequestBody EventInquiryDto eventInquiryDto, @RequestHeader String authorization) {
         LOGGER.info("POST /api/v1/events/{}", eventInquiryDto);
-
+        Profile profile = tokenService.getProfileFromHeader(authorization);
+        eventInquiryDto.setEmployer(EmployerDto.EmployerDtoBuilder.aEmployerDto().withProfileDto(ProfileDto.ProfileDtoBuilder.aProfileDto().withId(profile.getId()).build()).build());
         return eventMapper.eventToEventInquiryDto(
             eventService.saveEvent(eventMapper.eventInquiryDtoToEvent(eventInquiryDto)));
 
@@ -64,28 +74,19 @@ public class EventEndpoint {
     @ResponseBody
     public List<DetailedEventDto> search(SearchEventDto searchEventDto) {
         LOGGER.info("GET /api/v1/events");
-        System.out.println(searchEventDto);
         return eventMapper.eventsToDetailedEventDtos(eventService.findAll(searchEventDto));
-    }
-
-
-    @PutMapping
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @ApiOperation(value = "Update an event", authorizations = {@Authorization(value = "apiKey")})
-    @PreAuthorize("hasAuthority('ROLE_EMPLOYER')")
-    public void update(@Valid @RequestBody EventInquiryDto eventInquiryDto) {
-        LOGGER.info("PUT /api/v1/events/{}", eventInquiryDto);
-        eventMapper.eventToEventInquiryDto(
-            eventService.saveEvent(eventMapper.eventInquiryDtoToEvent(eventInquiryDto)));
-
     }
 
     @DeleteMapping(value = "/{id}")
     @ResponseStatus(HttpStatus.OK)
     @ApiOperation(value = "Delete event and related data")
     @PreAuthorize("hasAuthority('ROLE_EMPLOYER')")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Long id, @RequestHeader String authorization) {
         LOGGER.info("DELETE /api/v1/events/{}", id);
+        Event e = eventService.findById(id);
+        if(!tokenService.getProfileFromHeader(authorization).getId().equals(e.getEmployer().getId())) {
+            throw new AuthorizationServiceException(String.format("Keine Berchtigung um das Event \"%s\" zu löschen", e.getTitle()));
+        }
         eventService.deleteEventById(id);
     }
